Move the transaction types to a new module
The types moved are:

- `GenericTransaction`
- `TransactionInfo`
- `TransactionSigned`
- `TransactionUnsigned`
- `Transaction1559Signed`
- `Transaction1559Unsigned`
- `Transaction2930Signed`
- `Transaction2930Unsigned`
- `Transaction4844Signed`
- `Transaction4844Unsigned`
- `Transaction7702Signed`
- `Transaction7702Unsigned`
- `TransactionLegacySigned`
- `TransactionLegacyUnsigned`
- `AccessList`
- `AccessListEntry`
- `AuthorizationListEntry`
- `HashesOrTransactionInfos`
- `InputOrData`
