+++ substrate/frame/revive/src/evm/api/transaction.rs
@@ -1,0 +1,710 @@
+// This file is part of Substrate.
+
+// Copyright (C) Parity Technologies (UK) Ltd.
+// SPDX-License-Identifier: Apache-2.0
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#![allow(missing_docs)]
+
+use super::{Byte, Bytes, TypeEip1559, TypeEip2930, TypeEip4844, TypeEip7702, TypeLegacy};
+use codec::{Decode, DecodeWithMemTracking, Encode};
+use derive_more::{From, TryInto};
+use ethereum_types::*;
+use scale_info::TypeInfo;
+use serde::{Deserialize, Deserializer, Serialize};
+
+/// Transaction object generic to all types
+#[derive(
+	Debug, Default, Clone, Encode, Decode, TypeInfo, Serialize, Deserialize, Eq, PartialEq,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct GenericTransaction {
+	/// accessList
+	/// EIP-2930 access list
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub access_list: Option<AccessList>,
+	/// authorizationList
+	/// List of account code authorizations (EIP-7702)
+	#[serde(default, skip_serializing_if = "Vec::is_empty")]
+	pub authorization_list: Vec<AuthorizationListEntry>,
+	/// blobVersionedHashes
+	/// List of versioned blob hashes associated with the transaction's EIP-4844 data blobs.
+	#[serde(default)]
+	pub blob_versioned_hashes: Vec<H256>,
+	/// blobs
+	/// Raw blob data.
+	#[serde(default, skip_serializing_if = "Vec::is_empty")]
+	pub blobs: Vec<Bytes>,
+	/// chainId
+	/// Chain ID that this transaction is valid on.
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub chain_id: Option<U256>,
+	/// from address
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub from: Option<Address>,
+	/// gas limit
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub gas: Option<U256>,
+	/// gas price
+	/// The gas price willing to be paid by the sender in wei
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub gas_price: Option<U256>,
+	/// input data
+	#[serde(flatten, deserialize_with = "deserialize_input_or_data")]
+	pub input: InputOrData,
+	/// max fee per blob gas
+	/// The maximum total fee per gas the sender is willing to pay for blob gas in wei
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub max_fee_per_blob_gas: Option<U256>,
+	/// max fee per gas
+	/// The maximum total fee per gas the sender is willing to pay (includes the network / base fee
+	/// and miner / priority fee) in wei
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub max_fee_per_gas: Option<U256>,
+	/// max priority fee per gas
+	/// Maximum fee per gas the sender is willing to pay to miners in wei
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub max_priority_fee_per_gas: Option<U256>,
+	/// nonce
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub nonce: Option<U256>,
+	/// to address
+	pub to: Option<Address>,
+	/// type
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub r#type: Option<Byte>,
+	/// value
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub value: Option<U256>,
+}
+
+/// Transaction information
+#[derive(Debug, Default, Clone, Serialize, Eq, PartialEq, TypeInfo, Encode, Decode)]
+#[serde(rename_all = "camelCase")]
+pub struct TransactionInfo {
+	/// block hash
+	pub block_hash: H256,
+	/// block number
+	pub block_number: U256,
+	/// from address
+	pub from: Address,
+	/// transaction hash
+	pub hash: H256,
+	/// transaction index
+	pub transaction_index: U256,
+	#[serde(flatten)]
+	pub transaction_signed: TransactionSigned,
+}
+
+// Custom deserializer to work around serde's limitation with flatten + untagged enums from Value
+// See: https://github.com/serde-rs/serde/issues/1183
+impl<'de> Deserialize<'de> for TransactionInfo {
+	fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+	where
+		D: serde::Deserializer<'de>,
+	{
+		use alloc::{collections::BTreeMap, string::String};
+		use serde::de::Error;
+
+		// First try deserializing to a map
+		let mut map = <BTreeMap<String, serde_json::Value>>::deserialize(deserializer)?;
+
+		// Extract the TransactionInfo-specific fields
+		let block_hash =
+			map.remove("blockHash").ok_or_else(|| D::Error::missing_field("blockHash"))?;
+		let block_number = map
+			.remove("blockNumber")
+			.ok_or_else(|| D::Error::missing_field("blockNumber"))?;
+		let from = map.remove("from").ok_or_else(|| D::Error::missing_field("from"))?;
+		let hash = map.remove("hash").ok_or_else(|| D::Error::missing_field("hash"))?;
+		let transaction_index = map
+			.remove("transactionIndex")
+			.ok_or_else(|| D::Error::missing_field("transactionIndex"))?;
+
+		// The remaining fields should be for TransactionSigned
+		// Convert back to JSON and deserialize
+		let remaining = serde_json::Value::Object(map.into_iter().collect());
+		let json_str = serde_json::to_string(&remaining).map_err(D::Error::custom)?;
+		let transaction_signed: TransactionSigned =
+			serde_json::from_str(&json_str).map_err(D::Error::custom)?;
+
+		Ok(Self {
+			block_hash: serde_json::from_value(block_hash).map_err(D::Error::custom)?,
+			block_number: serde_json::from_value(block_number).map_err(D::Error::custom)?,
+			from: serde_json::from_value(from).map_err(D::Error::custom)?,
+			hash: serde_json::from_value(hash).map_err(D::Error::custom)?,
+			transaction_index: serde_json::from_value(transaction_index)
+				.map_err(D::Error::custom)?,
+			transaction_signed,
+		})
+	}
+}
+
+#[derive(
+	Debug,
+	Clone,
+	Serialize,
+	Deserialize,
+	From,
+	TryInto,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(untagged)]
+pub enum TransactionSigned {
+	Transaction7702Signed(Transaction7702Signed),
+	Transaction4844Signed(Transaction4844Signed),
+	Transaction1559Signed(Transaction1559Signed),
+	Transaction2930Signed(Transaction2930Signed),
+	TransactionLegacySigned(TransactionLegacySigned),
+}
+
+impl Default for TransactionSigned {
+	fn default() -> Self {
+		TransactionSigned::TransactionLegacySigned(Default::default())
+	}
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, From, TryInto, Eq, PartialEq)]
+#[serde(untagged)]
+pub enum TransactionUnsigned {
+	Transaction7702Unsigned(Transaction7702Unsigned),
+	Transaction4844Unsigned(Transaction4844Unsigned),
+	Transaction1559Unsigned(Transaction1559Unsigned),
+	Transaction2930Unsigned(Transaction2930Unsigned),
+	TransactionLegacyUnsigned(TransactionLegacyUnsigned),
+}
+
+impl Default for TransactionUnsigned {
+	fn default() -> Self {
+		TransactionUnsigned::TransactionLegacyUnsigned(Default::default())
+	}
+}
+
+/// EIP-1559 transaction.
+#[derive(
+	Debug,
+	Default,
+	Clone,
+	Serialize,
+	Deserialize,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct Transaction1559Unsigned {
+	/// accessList
+	/// EIP-2930 access list
+	pub access_list: AccessList,
+	/// chainId
+	/// Chain ID that this transaction is valid on.
+	pub chain_id: U256,
+	/// gas limit
+	pub gas: U256,
+	/// gas price
+	/// The effective gas price paid by the sender in wei. For transactions not yet included in a
+	/// block, this value should be set equal to the max fee per gas. This field is DEPRECATED,
+	/// please transition to using effectiveGasPrice in the receipt object going forward.
+	pub gas_price: U256,
+	/// input data
+	pub input: Bytes,
+	/// max fee per gas
+	/// The maximum total fee per gas the sender is willing to pay (includes the network / base fee
+	/// and miner / priority fee) in wei
+	pub max_fee_per_gas: U256,
+	/// max priority fee per gas
+	/// Maximum fee per gas the sender is willing to pay to miners in wei
+	pub max_priority_fee_per_gas: U256,
+	/// nonce
+	pub nonce: U256,
+	/// to address
+	pub to: Option<Address>,
+	/// type
+	pub r#type: TypeEip1559,
+	/// value
+	pub value: U256,
+}
+/// EIP-2930 transaction.
+#[derive(
+	Debug,
+	Default,
+	Clone,
+	Serialize,
+	Deserialize,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct Transaction2930Unsigned {
+	/// accessList
+	/// EIP-2930 access list
+	pub access_list: AccessList,
+	/// chainId
+	/// Chain ID that this transaction is valid on.
+	pub chain_id: U256,
+	/// gas limit
+	pub gas: U256,
+	/// gas price
+	/// The gas price willing to be paid by the sender in wei
+	pub gas_price: U256,
+	/// input data
+	pub input: Bytes,
+	/// nonce
+	pub nonce: U256,
+	/// to address
+	pub to: Option<Address>,
+	/// type
+	pub r#type: TypeEip2930,
+	/// value
+	pub value: U256,
+}
+/// EIP-4844 transaction.
+#[derive(
+	Debug,
+	Default,
+	Clone,
+	Serialize,
+	Deserialize,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct Transaction4844Unsigned {
+	/// accessList
+	/// EIP-2930 access list
+	pub access_list: AccessList,
+	/// blobVersionedHashes
+	/// List of versioned blob hashes associated with the transaction's EIP-4844 data blobs.
+	pub blob_versioned_hashes: Vec<H256>,
+	/// chainId
+	/// Chain ID that this transaction is valid on.
+	pub chain_id: U256,
+	/// gas limit
+	pub gas: U256,
+	/// input data
+	pub input: Bytes,
+	/// max fee per blob gas
+	/// The maximum total fee per gas the sender is willing to pay for blob gas in wei
+	pub max_fee_per_blob_gas: U256,
+	/// max fee per gas
+	/// The maximum total fee per gas the sender is willing to pay (includes the network / base fee
+	/// and miner / priority fee) in wei
+	pub max_fee_per_gas: U256,
+	/// max priority fee per gas
+	/// Maximum fee per gas the sender is willing to pay to miners in wei
+	pub max_priority_fee_per_gas: U256,
+	/// nonce
+	pub nonce: U256,
+	/// to address
+	pub to: Address,
+	/// type
+	pub r#type: TypeEip4844,
+	/// value
+	pub value: U256,
+}
+/// Legacy transaction.
+#[derive(
+	Debug,
+	Default,
+	Clone,
+	Serialize,
+	Deserialize,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct TransactionLegacyUnsigned {
+	/// chainId
+	/// Chain ID that this transaction is valid on.
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub chain_id: Option<U256>,
+	/// gas limit
+	pub gas: U256,
+	/// gas price
+	/// The gas price willing to be paid by the sender in wei
+	pub gas_price: U256,
+	/// input data
+	pub input: Bytes,
+	/// nonce
+	pub nonce: U256,
+	/// to address
+	pub to: Option<Address>,
+	/// type
+	pub r#type: TypeLegacy,
+	/// value
+	pub value: U256,
+}
+/// EIP-7702 transaction.
+#[derive(
+	Debug,
+	Clone,
+	Serialize,
+	Deserialize,
+	Default,
+	From,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct Transaction7702Unsigned {
+	/// accessList
+	/// EIP-2930 access list
+	pub access_list: AccessList,
+	/// authorizationList
+	/// List of account code authorizations
+	pub authorization_list: Vec<AuthorizationListEntry>,
+	/// chainId
+	/// Chain ID that this transaction is valid on.
+	pub chain_id: U256,
+	/// gas limit
+	pub gas: U256,
+	/// input data
+	pub input: Bytes,
+	/// max fee per gas
+	/// The maximum total fee per gas the sender is willing to pay (includes the network / base fee
+	/// and miner / priority fee) in wei
+	pub max_fee_per_gas: U256,
+	/// max priority fee per gas
+	/// Maximum fee per gas the sender is willing to pay to miners in wei
+	pub max_priority_fee_per_gas: U256,
+	/// nonce
+	pub nonce: U256,
+	/// to address
+	///
+	/// # Note
+	///
+	/// Extracted from eip-7702: `Note, this implies a null destination is not valid.`
+	pub to: Address,
+	/// type
+	pub r#type: TypeEip7702,
+	/// value
+	pub value: U256,
+}
+/// Signed 7702 Transaction
+#[derive(
+	Debug,
+	Clone,
+	Serialize,
+	Deserialize,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct Transaction7702Signed {
+	#[serde(flatten)]
+	pub transaction_7702_unsigned: Transaction7702Unsigned,
+	/// r
+	pub r: U256,
+	/// s
+	pub s: U256,
+	/// v
+	/// For backwards compatibility, `v` is optionally provided as an alternative to `yParity`.
+	/// This field is DEPRECATED and all use of it should migrate to `yParity`.
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub v: Option<U256>,
+	/// yParity
+	/// The parity (0 for even, 1 for odd) of the y-value of the secp256k1 signature.
+	pub y_parity: U256,
+}
+/// Signed 1559 Transaction
+#[derive(
+	Debug,
+	Default,
+	Clone,
+	Serialize,
+	Deserialize,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct Transaction1559Signed {
+	#[serde(flatten)]
+	pub transaction_1559_unsigned: Transaction1559Unsigned,
+	/// r
+	pub r: U256,
+	/// s
+	pub s: U256,
+	/// v
+	/// For backwards compatibility, `v` is optionally provided as an alternative to `yParity`.
+	/// This field is DEPRECATED and all use of it should migrate to `yParity`.
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub v: Option<U256>,
+	/// yParity
+	/// The parity (0 for even, 1 for odd) of the y-value of the secp256k1 signature.
+	pub y_parity: U256,
+}
+/// Signed 2930 Transaction
+#[derive(
+	Debug,
+	Default,
+	Clone,
+	Serialize,
+	Deserialize,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct Transaction2930Signed {
+	#[serde(flatten)]
+	pub transaction_2930_unsigned: Transaction2930Unsigned,
+	/// r
+	pub r: U256,
+	/// s
+	pub s: U256,
+	/// v
+	/// For backwards compatibility, `v` is optionally provided as an alternative to `yParity`.
+	/// This field is DEPRECATED and all use of it should migrate to `yParity`.
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub v: Option<U256>,
+	/// yParity
+	/// The parity (0 for even, 1 for odd) of the y-value of the secp256k1 signature.
+	pub y_parity: U256,
+}
+/// Signed 4844 Transaction
+#[derive(
+	Debug,
+	Default,
+	Clone,
+	Serialize,
+	Deserialize,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct Transaction4844Signed {
+	#[serde(flatten)]
+	pub transaction_4844_unsigned: Transaction4844Unsigned,
+	/// r
+	pub r: U256,
+	/// s
+	pub s: U256,
+	/// yParity
+	/// The parity (0 for even, 1 for odd) of the y-value of the secp256k1 signature.
+	pub y_parity: U256,
+}
+/// Signed Legacy Transaction
+#[derive(
+	Debug,
+	Default,
+	Clone,
+	Serialize,
+	Deserialize,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct TransactionLegacySigned {
+	#[serde(flatten)]
+	pub transaction_legacy_unsigned: TransactionLegacyUnsigned,
+	/// r
+	pub r: U256,
+	/// s
+	pub s: U256,
+	/// v
+	pub v: U256,
+}
+
+/// Access list
+pub type AccessList = Vec<AccessListEntry>;
+
+/// Access list entry
+#[derive(
+	Debug,
+	Default,
+	Clone,
+	Encode,
+	Decode,
+	TypeInfo,
+	Serialize,
+	Deserialize,
+	Eq,
+	PartialEq,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct AccessListEntry {
+	pub address: Address,
+	pub storage_keys: Vec<H256>,
+}
+
+/// Authorization list entry for EIP-7702
+#[derive(
+	Debug,
+	Default,
+	Clone,
+	Serialize,
+	Deserialize,
+	Eq,
+	PartialEq,
+	TypeInfo,
+	Encode,
+	Decode,
+	DecodeWithMemTracking,
+)]
+#[serde(rename_all = "camelCase")]
+pub struct AuthorizationListEntry {
+	/// Chain ID that this authorization is valid on
+	pub chain_id: U256,
+	/// Address to authorize
+	pub address: Address,
+	/// Nonce of the authorization
+	pub nonce: U256,
+	/// y-parity of the signature
+	pub y_parity: U256,
+	/// r component of signature
+	pub r: U256,
+	/// s component of signature
+	pub s: U256,
+}
+
+#[derive(
+	Debug, Clone, Serialize, Deserialize, From, TryInto, Eq, PartialEq, TypeInfo, Encode, Decode,
+)]
+#[serde(untagged)]
+pub enum HashesOrTransactionInfos {
+	/// Transaction hashes
+	Hashes(Vec<H256>),
+	/// Full transactions
+	TransactionInfos(Vec<TransactionInfo>),
+}
+
+impl Default for HashesOrTransactionInfos {
+	fn default() -> Self {
+		HashesOrTransactionInfos::Hashes(Default::default())
+	}
+}
+
+impl HashesOrTransactionInfos {
+	pub fn push_hash(&mut self, hash: H256) {
+		match self {
+			HashesOrTransactionInfos::Hashes(hashes) => hashes.push(hash),
+			_ => {},
+		}
+	}
+
+	pub fn len(&self) -> usize {
+		match self {
+			HashesOrTransactionInfos::Hashes(v) => v.len(),
+			HashesOrTransactionInfos::TransactionInfos(v) => v.len(),
+		}
+	}
+
+	pub fn is_empty(&self) -> bool {
+		self.len() == 0
+	}
+
+	pub fn contains_tx(&self, hash: H256) -> bool {
+		match self {
+			HashesOrTransactionInfos::Hashes(hashes) => hashes.iter().any(|h256| *h256 == hash),
+			HashesOrTransactionInfos::TransactionInfos(transaction_infos) => {
+				transaction_infos.iter().any(|ti| ti.hash == hash)
+			},
+		}
+	}
+}
+
+/// Input of a `GenericTransaction`
+#[derive(
+	Debug, Default, Clone, Encode, Decode, TypeInfo, Serialize, Deserialize, Eq, PartialEq,
+)]
+pub struct InputOrData {
+	#[serde(skip_serializing_if = "Option::is_none")]
+	input: Option<Bytes>,
+	#[serde(skip_serializing_if = "Option::is_none")]
+	data: Option<Bytes>,
+}
+
+impl From<Bytes> for InputOrData {
+	fn from(value: Bytes) -> Self {
+		InputOrData { input: Some(value), data: None }
+	}
+}
+
+impl From<Vec<u8>> for InputOrData {
+	fn from(value: Vec<u8>) -> Self {
+		InputOrData { input: Some(Bytes(value)), data: None }
+	}
+}
+
+impl InputOrData {
+	/// Get the input as `Bytes`.
+	pub fn to_bytes(self) -> Bytes {
+		match self {
+			InputOrData { input: Some(input), data: _ } => input,
+			InputOrData { input: None, data: Some(data) } => data,
+			_ => Default::default(),
+		}
+	}
+
+	/// Get the input as `Vec<u8>`.
+	pub fn to_vec(self) -> Vec<u8> {
+		self.to_bytes().0
+	}
+}
+
+fn deserialize_input_or_data<'d, D: Deserializer<'d>>(d: D) -> Result<InputOrData, D::Error> {
+	let value = InputOrData::deserialize(d)?;
+	match &value {
+		InputOrData { input: Some(input), data: Some(data) } if input != data => {
+			Err(serde::de::Error::custom(
+				"Both \"data\" and \"input\" are set and not equal. Please use \"input\" to pass transaction call data",
+			))
+		},
+		_ => Ok(value),
+	}
+}
